refactorings
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@9149 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: 5947334066fcd8dfbcf60edf58d6daf3d1d02d9f

+++ scheduling/scheduler.slascheduler/src/de/uniol/inf/is/odysseus/scheduler/slascheduler/SLAConformanceFactory.java
@@ -35,7 +35,7 @@ public class SLAConformanceFactory {
 						plan);
 			} else if (sla.getScope() instanceof Number) {
 				conformance = new LatencyNumberConformance(distributor, sla, 
-						plan, (Long) sla.getMetric().getValue());
+						plan, sla.getMetric().getValue());
 			} else if (sla.getScope() instanceof Single) {
 				conformance = new LatencySingleConformance(distributor, sla, 
 						plan);
